requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
openpyxl==3.1.2
xlsxwriter==3.2.0
openai==1.12.0
schedule==1.2.1
playwright==1.41.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
selenium==4.18.0
webdriver-manager==4.0.1
retry==0.9.2
tenacity==8.2.3
